requirements.txt
asgiref==3.7.2
azure-core==1.29.5
azure-storage-blob==12.19.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
Django==4.2.7
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-taggit==5.0.1
idna==3.4
install==1.3.5
isodate==0.6.1
Pillow==10.1.0
pycparser==2.21
python-decouple==3.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
